Revise "used in production" question
Instead of voting on whether or not a project has documented that it is
being used in production, TAC voting members should vote on whether or
not a project is actually used in production. TAC voting members should
evaluate this to the best of their knowledge.

It is helpful for a project to provide evidence of production use in
their proposal. TAC voting members should vote on the truth of that
evidence. TAC voting members should not vote on whether or not the
evidence was provided by the project.

Signed-off-by: Justin LaPolla <justin.lapolla@hpe.com>