The TAC does not vote on project charters
The project charter is a Linux Foundation requirement that The Linux
Foundation will enact and enforce when they onboard the project. Some
projects do not have a charter yet when the TAC votes on their project
proposal. Lack of a charter should not prevent the TAC from
conditionally approving a project.

Signed-off-by: Justin LaPolla <justin.lapolla@hpe.com>