Draft TAC repo process document
I propose to just have it in the CONTRIBUTING file because it is a
widely recognized naming scheme on GH and in the end this is close
enough to what we try to describe here.
Let me know if you disagree and how what other filename would be better
in your opinion and why.

Signed-off-by: Damien L-G <dalg24@gmail.com>