Clarify vote quorum requirements
In the TAC meeting on December 11, 2025, we had a poll result where 8
votes were cast and there are 16 voting TAC members, but we did not
approve the resolution because of lack of quorum. It is not sufficient
for the number of votes cast to be equal to half the number of voting
TAC members. For a poll to meet the quorum requirement, the number of
votes cast needs to be greater than half the number of voting TAC
members.

Update the Project Lifecycle Policies to reflect that.

Signed-off-by: Justin LaPolla <justin.lapolla@hpe.com>